+++ app/(product)/logs/page.tsx
@@ -1,9 +1,8 @@
-"use client";
 import React from 'react';
 import AgentLogStream from '@/components/AgentLogStream';
 export const revalidate = 0 as const;
 export const dynamic = "force-dynamic";
-const AgentLogsPage: React.FC = () => {
+export default function AgentLogsPage() {
   return (
     <main className="min-h-screen bg-gray-50 p-6" suppressHydrationWarning>
       <header className="text-center mb-8">
@@ -13,7 +12,5 @@ const AgentLogsPage: React.FC = () => {
       <AgentLogStream />
     </main>
   );
-};
-
-export default AgentLogsPage;
+}
