+++ app/(product)/predictions/page.tsx
@@ -1,10 +1,7 @@
-"use client";
 import React from 'react';
 import MatchupInsights from '@/components/predictions/MatchupInsights';
 export const revalidate = 0 as const;
 export const dynamic = "force-dynamic";
-const PredictionsPage: React.FC = () => {
+export default function PredictionsPage() {
   return <MatchupInsights />;
-};
-
-export default PredictionsPage;
+}
