+++ app/[locale]/page.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useContext } from 'react';
 import { I18nContext } from '@/lib/i18n/config';
 import { LocaleSwitcher } from '@/lib/i18n/LocaleSwitcher';
