+++ app/agent-interface/page.tsx
@@ -1,25 +1,7 @@
-"use client";
-
-import { useState } from 'react';
-import nextDynamic from 'next/dynamic';
+import AgentInterface from './AgentInterfaceClient';
 export const revalidate = 0 as const;
 export const dynamic = "force-dynamic";
-const AgentFlowVisualizer = nextDynamic(() => import('@/components/AgentFlowVisualizer'), { ssr: false });
 
 export default function Page() {
-  const [show, setShow] = useState(false);
-  return (
-    <div className="p-4">
-      {show ? (
-        <AgentFlowVisualizer />
-      ) : (
-        <button
-          className="rounded border px-4 py-2"
-          onClick={() => setShow(true)}
-        >
-          Show Live Agent Network
-        </button>
-      )}
-    </div>
-  );
+  return <AgentInterface />;
 }
