+++ app/toast-demo/page.tsx
@@ -1,23 +1,7 @@
-'use client';
-
-import React, { useEffect } from 'react';
-import { toast } from '@/lib/ui/toast';
+import ToastDemoPage from './ToastDemoClient';
 export const revalidate = 0 as const;
 export const dynamic = "force-dynamic";
 
-export default function ToastDemoPage() {
-  useEffect(() => {
-    toast.success('Success toast');
-    toast.error('Error toast');
-    toast.promise(
-      new Promise((resolve) => setTimeout(resolve, 1000)),
-      {
-        loading: 'Loading toast',
-        success: 'Loaded toast',
-        error: 'Failed toast',
-      },
-    );
-  }, []);
-
-  return <p>Toast demo</p>;
+export default function Page() {
+  return <ToastDemoPage />;
 }
