+++ lib/config/env.js
@@ -8,8 +8,10 @@ let apiKey = process.env.SPORTS_API_KEY;
 if (!apiKey && provider === 'thesportsdb' && !isProdLike) {
   apiKey = '123';
 }
-if (isProdLike && !apiKey) {
-  throw new Error('[env] Missing required env: SPORTS_API_KEY');
+
+export function getSportsApiKey() {
+  if (!apiKey) throw new Error('[env] Missing required env: SPORTS_API_KEY');
+  return apiKey;
 }
 
 export const ENV = {
